backend/requirements.txt
fastapi[all]
starlette
uvicorn
boto3
requests
openai
faiss-cpu
numpy
pydantic
python-dotenv
